added sendCoin

+++ MinterWallet/Classes/Popup/ViewModel/SendPopupViewModel.swift
@@ -11,7 +11,7 @@ import UIKit
 
 class SendPopupViewModel : PopupViewModel {
 	
-	var amount: UInt?
+	var amount: Double?
 	
 	var coin: String?
